Refactor errors into classes
Each type of error is encoded in the name and signature of a virtual
error_reporter function:

    virtual void report_error_assignment_before_variable_declaration(
      identifier assignment, identifier declaration);

This formulation is not a problem, but for some reason, I like the
following style better:

    struct error_assignment_before_variable_declaration {
      identifier assignment;
      identifier declaration;
    };
    virtual void report(error_assignment_before_variable_declaration);

Refactor into the latter form: a class containing all of the parameters,
and an overloaded 'report' function.

This commit should not change behavior.