Use std::variant for errors in tests
error_collector has knowledge about every error type. This makes adding
new error types difficult; in addition to creating the type
itself, the emitter of the error, and the test for the error, you also
need to add a few entries in error_collector.

Refactor error_collector to use an std::variant instead of a manual
union. This means asserting code is uglier and harder to write, but
there are fewer places to change when adding a new error type.